% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fastpcdtest.R
\docType{package}
\name{fastpcdtest}
\alias{fastpcdtest}
\alias{fastpcdtest-package}
\title{fastpcdtest: A package for fast and robust calculation of Pesaran's CD test.}
\description{
The fastpcdtest package calculates Pesaran's CD test. It substitutes
for pcdtest in plm package in two situations:
1) when the panel is huges (since it is faster than pcdtest in plm package) and
2) when the panel is highly unbalanced (since is calculates CD test even when
some observations do not have overlap in their time dimension).
}
\section{fastpcdtest functions}{

The package provides only one function, \code{pcdtest}, which overshadows
the test of the same name in \bold{plm} package. However, internally, there are
two Rcpp functions, one for balanced panels and other for unbalanced panels.
R wrapper function \code{pcdtest} automatically selects the proper function.
}

