% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fastpcdtest.R
\name{pcdtest}
\alias{pcdtest}
\title{Calculate Pesaran's CD test.}
\usage{
pcdtest(mod, index = NULL, min_common = 2L)
}
\arguments{
\item{mod}{Either a plm model or data.frame with specified index.}

\item{index}{Character vector of length three containing names of columns with
cross-section unit (1st position), time unit (2nd position) and
residuals (3rd position).}

\item{min_common}{}
}
\value{
Object of htest class with the results of the test.
}
\description{
description
}
\details{
details
}
\examples{
#to be added
}

